+++ source/backend/Risk.API/Models/RequestBodies/CambiarEstadoMensajeriaRequestBody.cs
@@ -27,7 +27,7 @@ namespace Risk.API.Models
     public class CambiarEstadoMensajeriaRequestBody
     {
         public TipoMensajeria TipoMensajeria { get; set; }
-        public int IdMensajeria { get; set; }
+        public long IdMensajeria { get; set; }
         public EstadoMensajeria Estado { get; set; }
         public string RespuestaEnvio { get; set; }
     }
